requirements.txt
psycopg2
pyyaml
regex
tiktoken
openai
langchain_core
langchain_openai
langgraph
uvicorn